fix: 删除新增的互助码js，复用原始jd_get_share_code.js
